services/gather/requirements.txt
flask[async]==2.2.3
pandas==1.5.3
python-dotenv==1.0.0
nest-asyncio==1.5.6
aiohttp==3.8.4
Pillow==8.4.0